+++ src/main/java/subway/domain/subway/route_map/RouteMap.java
@@ -1,4 +1,4 @@
-package subway.domain.subway.routeMap;
+package subway.domain.subway.route_map;
 
 import subway.domain.Path;
 import subway.domain.station.Station;
