+++ src/test/java/subway/domain/subway/billing_policy/BillingPolicyByDistanceTest.java
@@ -1,4 +1,4 @@
-package subway.domain.subway;
+package subway.domain.subway.billing_policy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import subway.domain.Path;
-import subway.domain.subway.billing_policy.BillingPolicyByDistance;
-import subway.domain.subway.billing_policy.Fare;
 
 class BillingPolicyByDistanceTest {
